# EastsydApp.API/bin/Release/net7.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "EastsydApp.API, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
...
    "/api/Product/{id}": {
      "get": {
        "tags": [
          "Product"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/ProductDto"
            }
          }
        }
      }
    },
...
    "/api/Product/{categoryId}/GetItemsByCategory": {
      "get": {
        "tags": [
          "Product"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "categoryId",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/ProductDto"
              }
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "ProductCategoryDto": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "name": {
          "type": "string"
        },
        "iconCss": {
          "type": "string"
        },
        "products": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ProductDto"
          }
        }
      },
      "additionalProperties": false
    },
    "ProductDto": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "name": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "imageURL": {
          "type": "string"
        },
        "price": {
          "format": "double",
          "type": "number"
        },
        "qty": {
          "format": "int32",
          "type": "integer"
        },
        "productCategoryId": {
          "format": "int32",
          "type": "integer"
        },
        "categoryName": {
          "type": "string"
        },
        "productCategory": {
          "$ref": "#/definitions/ProductCategoryDto"
        }
      },
      "additionalProperties": false
    },
...
  }
}